Complete major organization and comments refactor
