+++ avalanche/networks/network_genesis_config.go
