--- commons/ava_testsuite/avalanche_tester.go
@@ -1,5 +1,0 @@
-package ava_testsuite
-
-type AvalancheTester interface {
-	ExecuteTest() error
-}
