+++ gecko_client/utils/constants/chain_constants.go
@@ -11,10 +11,13 @@ var (
 	PlatformChainID ids.ID
 	// CChainID ...
 	CChainID ids.ID
+	// AvaxAssetID ...
+	AvaxAssetID ids.ID
 )
 
 func init() {
 	XChainID, _ = ids.FromString("2VvmkRw4yrz8tPrVnCCbvEK1JxNyujpqhmU6SGonxMpkWBx9UD")
 	PlatformChainID = ids.Empty
 	CChainID, _ = ids.FromString("f5DjTrC9YJPagt9ogKgKPYpp7KMaCBKsv7AeqfonpTiw6rBec")
+	AvaxAssetID, _ = ids.FromString("2TrXx5kLGWa9RP3RiYWi7VkmNbppwPU4DCmTdqwuKzGFE7fsvP")
 }
