+++ testsuite/tester/avalanche_tester.go
@@ -1,0 +1,6 @@
+package tester
+
+// AvalancheTester is the interface for a ready to execute test
+type AvalancheTester interface {
+	ExecuteTest() error
+}
